requirements.txt
fastapi
uvicorn
httpx
psycopg2-binary
aiogram==2.25.1
python-dotenv
aiohttp